requirements.txt
flask
pydantic
PyYAML
requests
